Task2/web_services_app/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
gunicorn==21.2.0
pymongo==4.6.1
python-dotenv==1.0.1
